--- api/config.py
@@ -1,25 +1,0 @@
-prod_url = "https://durger-king-five.vercel.app"
-test_url = "https://durger-king-daviesdev.vercel.app"
-url = prod_url
-packages = {
-    "250": {
-        "w": "30",
-        "m": "80"
-    },
-    "1000": {
-        "w": "60",
-        "m": "150"
-    },  
-    "1500": {
-        "w": "80",
-        "m": "200"
-    },
-    "2500": {
-        "w": "120",
-        "m": "300"
-    },
-    "5000": {
-        "w": "180",
-        "m": "500"
-    }
-}
