+++ config.py
@@ -1,0 +1,25 @@
+prod_url = "https://julesfoundationschool.org"
+test_url = "https://"
+url = prod_url
+packages = {
+    "250": {
+        "w": "30",
+        "m": "80"
+    },
+    "1000": {
+        "w": "60",
+        "m": "150"
+    },  
+    "1500": {
+        "w": "80",
+        "m": "200"
+    },
+    "2500": {
+        "w": "120",
+        "m": "300"
+    },
+    "5000": {
+        "w": "180",
+        "m": "500"
+    }
+}
