--- vercel.json
@@ -1,5 +1,0 @@
-{
-    "rewrites": [
-      { "source": "/(.*)", "destination": "/api/index" }
-    ]
-  }
